+++ src/example_app/example_app/documents/resources.py
@@ -3,15 +3,11 @@
 :Authors: cykooz
 :Date: 31.07.2020
 """
-from dataclasses import dataclass, asdict
+from dataclasses import dataclass
 from typing import Dict
 
 from pyramid.security import Allow, Everyone
-from restfw.hal import HalResource, HalResourceWithEmbedded, list_to_embedded_resources
-from restfw.interfaces import MethodOptions
-from restfw.schemas import GetEmbeddedSchema
-
-from . import schemas
+from restfw.hal import HalResource
 
 
 @dataclass()
@@ -29,29 +25,18 @@ class Doc(HalResource):
         self.__parent__ = parent
         self.__name__ = str(model.id)
 
-    options_for_get = MethodOptions(None, schemas.DocSchema, permission='docs.get')
-
-    def as_dict(self, request):
-        return asdict(self.model)
-
-    options_for_patch = MethodOptions(schemas.PatchDocSchema, schemas.DocSchema,
-                                      permission='docs.edit')
-
     def http_patch(self, request, params):
         for key, value in params.items():
             if hasattr(self.model, key):
                 setattr(self.model, key, value)
         created = False
-        return self, created
-
-    options_for_delete = MethodOptions(None, None, permission='docs.edit')
+        return created
 
     def http_delete(self, request, params):
         self.__parent__.delete_doc(self.model.id)
 
 
-class Docs(HalResourceWithEmbedded):
-
+class Docs(HalResource):
     __acl__ = [
         # (Allow, Authenticated, 'docs.'),
         # DENY_ALL,
@@ -59,33 +44,19 @@ class Docs(HalResourceWithEmbedded):
     ]
 
     def __init__(self):
-        self._models: Dict[str, DocModel] = {}
+        self.models: Dict[str, DocModel] = {}
         self._next_id = 1
 
     def __getitem__(self, key):
-        model = self._models.get(key)
+        model = self.models.get(key)
         if model:
             return Doc(model, parent=self)
         return super().__getitem__(key)
 
-    options_for_get = MethodOptions(GetEmbeddedSchema, schemas.DocsSchema,
-                                    permission='docs.get')
-
-    def get_embedded(self, request, params):
-        docs = [self.get_doc_by_model(model) for model in self._models.values()]
-        return list_to_embedded_resources(
-            request, params, docs,
-            parent=self,
-            embedded_name='docs',
-        )
-
-    options_for_post = MethodOptions(schemas.CreateDocSchema, schemas.DocSchema,
-                                     permission='docs.edit')
-
     def http_post(self, request, params):
         model = DocModel(id=self._next_id, **params)
         self._next_id += 1
-        self._models[str(model.id)] = model
+        self.models[str(model.id)] = model
         resource = self.get_doc_by_model(model)
         created = True
         return resource, created
@@ -94,4 +65,4 @@ class Docs(HalResourceWithEmbedded):
         return Doc(model, parent=self)
 
     def delete_doc(self, doc_id: int):
-        self._models.pop(str(doc_id))
+        self.models.pop(str(doc_id))