+++ src/example_app/example_app/users/resources.py
@@ -3,16 +3,12 @@
 :Authors: cykooz
 :Date: 05.02.2020
 """
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 from typing import Dict, List, Literal, Optional, Union
 
 import pendulum
 from pyramid.security import Allow, Everyone
-from restfw.hal import HalResource, HalResourceWithEmbedded, list_to_embedded_resources
-from restfw.interfaces import MethodOptions
-from restfw.schemas import GetEmbeddedSchema
-
-from . import schemas
+from restfw.hal import HalResource
 
 
 @dataclass()
@@ -48,31 +44,17 @@ class User(HalResource):
         self.__parent__ = parent
         self.__name__ = str(model.id)
 
-    options_for_get = MethodOptions(None, schemas.UserSchema,
-                                    permission='users.get')
-
-    def as_dict(self, request):
-        res = asdict(self.model)
-        res['created'] = res['created'].isoformat()
-        return res
-
-    options_for_patch = MethodOptions(schemas.PatchUserSchema, schemas.UserSchema,
-                                      permission='users.edit')
-
     def http_patch(self, request, params: dict):
         for key, value in params.items():
             setattr(self.model, key, value)
         created = False
-        return self, created
-
-    options_for_delete = MethodOptions(None, None, permission='users.edit')
+        return created
 
     def http_delete(self, request, params):
         self.__parent__.delete_user(self.model.id)
 
 
-class Users(HalResourceWithEmbedded):
-
+class Users(HalResource):
     __acl__ = [
         # (Allow, Authenticated, 'users.'),
         # DENY_ALL,
@@ -80,29 +62,14 @@ class Users(HalResourceWithEmbedded):
     ]
 
     def __init__(self):
-        self._models: Dict[str, UserModel] = {}
+        self.models: Dict[str, UserModel] = {}
         self._next_id = 1
 
     def __getitem__(self, key):
-        model = self._models.get(key)
+        model = self.models.get(key)
         if model:
             return self.get_user_by_model(model)
         return super().__getitem__(key)
-
-    options_for_get = MethodOptions(GetEmbeddedSchema, schemas.UsersSchema,
-                                    permission='users.get')
-
-    def get_embedded(self, request, params):
-        users = [self.get_user_by_model(model) for model in self._models.values()]
-        return list_to_embedded_resources(
-            request, params,
-            resources=users,
-            parent=self,
-            embedded_name='users',
-        )
-
-    options_for_post = MethodOptions(schemas.CreateUserSchema, schemas.UserSchema,
-                                     permission='users.edit')
 
     def http_post(self, request, params):
         created = True
@@ -135,11 +102,11 @@ class Users(HalResourceWithEmbedded):
             current_work=current_work,
         )
         self._next_id += 1
-        self._models[str(user_id)] = model
+        self.models[str(user_id)] = model
         return self.get_user_by_model(model)
 
     def get_user_by_name(self, first_name):
-        for model in self._models.values():
+        for model in self.models.values():
             if model.first_name == first_name:
                 return self.get_user_by_model(model)
 
@@ -148,7 +115,7 @@ class Users(HalResourceWithEmbedded):
 
 
 def check_credentials(username, password, request):
-    users = request.root['users']  # type: Users
+    users: Users = request.root['users']
     user = users.get_user_by_name(username)
     if user:
         return [user.model.id]