+++ src/restfw_admin/restfw_admin/views.py
@@ -3,11 +3,18 @@
 :Authors: cykooz
 :Date: 18.07.2020
 """
+import dataclasses
+
 from pyramid.httpexceptions import HTTPMovedPermanently, HTTPOk
 from pyramid.request import Request
+from restfw import views
+from restfw.interfaces import MethodOptions
+from restfw.typing import PyramidRequest
 
+from . import schemas
 from .config import get_admin_ui_settings
-from .resources import get_admin
+from .models import ApiInfoModel
+from .resources import Admin, AdminChoices, ApiInfo, get_admin
 
 
 TEMPLATE = '''<script>
@@ -47,3 +54,55 @@ def admin_ui(request: Request):
 def redirect_to_admin_ui(request: Request):
     url = request.route_url('admin_ui_ts')
     return HTTPMovedPermanently(location=url)
+
+
+# ApiInfo
+
+@views.resource_view_config(ApiInfo)
+class ApiInfoView(views.HalResourceView):
+    resource: ApiInfo
+    options_for_get = MethodOptions(None, None, permission='rest_admin.api_info.get')
+
+    def as_dict(self):
+        registry = self.request.registry
+        title = registry.settings.get('restfw_admin.title', 'Admin UI')
+        root_url = registry.settings.get('restfw_admin.root_url', '')
+        if not root_url:
+            root_url = self.request.resource_url(self.request.root)
+        model = ApiInfoModel(
+            root_url=root_url.rstrip('/'),
+            title=title,
+            resources=self.resource.get_resources_info(self.request),
+        )
+        return dataclasses.asdict(model)
+
+
+# AdminChoices
+
+@views.resource_view_config(AdminChoices)
+class AdminChoicesView(views.HalResourceWithEmbeddedView):
+    resource: AdminChoices
+    options_for_get = MethodOptions(
+        schemas.GetAdminChoicesSchema, schemas.AdminChoicesSchema,
+        permission='admin_choices.get',
+    )
+
+    def get_embedded(self, params):
+        group = params.get('group')
+        choice_ids = params.get('id')
+        choices = list(self.resource.get_choices(self.request.registry, group, choice_ids))
+        return views.list_to_embedded_resources(
+            self.request, params, choices,
+            parent=self.resource,
+            embedded_name='choices',
+        )
+
+
+# Admin
+
+@views.resource_view_config(Admin)
+class AdminView(views.HalResourceView):
+
+    def http_get(self):
+        url = self.request.route_url('admin_ui_ts')
+        return HTTPMovedPermanently(location=url)